Added core module for angular app
